+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Experimental endpoint for cross-border available capacity (`/balancing/cross-border/available-capacity`)
+
 ## [1.15.0] - 2026-06-10
 
 ### Added